Error Handler system reworked.
There has been a major rework of the way SKY reports errors. It will not
longer write them to the /log/error dir. Every error will be written to
the core.log file and depending on the ENV, it will display it in HTML.
@todo: Figure out AJAX calls. @todo: Remove /log/error dir. @todo: Need
to add user error handeling in the core classes.